src/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django<6.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework-simplejwt
gunicorn==25.1.0
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
redis==6.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
waitress
django-cors-headers
stripe
django-jazzmin
# xhtml2pdf  <-- Temporarily removed to unblock build